:art: Add editorconfig
