--- src/server/api/models/popularBreed.ts
@@ -1,24 +1,0 @@
-import mongoose from 'mongoose';
-
-export const popularBreed = new mongoose.Schema({
-    _id: {
-        required: true,
-        type: String,
-    },
-    name: {
-        required: true,
-        type: String,
-    },
-    img_url: {
-        required: true,
-        type: String,
-    },
-    description: {
-        required: true,
-        type: String,
-    },
-    search_count: {
-        type: Number,
-        required: true,
-    },
-});
